[PLAT-8092] Customer password is visible in state file
Summary:
Customer password is required during creation of customer resource. This can be fetched from the environment variables like credentials for storage configs and cloud providers.
The validation for env variables is required only when the customer resource does not exist. Verified that password is not required in further steps.
Few lint changes in config files, and removed password as a field for customer resource.

Test Plan:
Terraform workflow

State file content for customer resource
```
{
  "api_token": "a6af6cc4-1ffd-402a-aad2-3f373790c25d",
  "code": "admin",
  "cuuid": "fe6f1848-e5e9-4e9d-a1cf-a0573f306a4f",
  "email": "demo@yugabyte.com",
  "id": "fe6f1848-e5e9-4e9d-a1cf-a0573f306a4f",
  "name": "demo",
  "timeouts": null
},
```

Reviewers: sneelakantan

Reviewed By: sneelakantan

Differential Revision: https://phabricator.dev.yugabyte.com/D24097