[PLAT-8957]Update module name in source code to be consistent with repository name
Summary:
The initial repository was called `terraform-provider-yugabyte-platform` and the same was used as
module name in the entire codebase. Currently for ease of coding, the repository and provider name
have been changed to `terraform-provider-yba` and `yba` respectively, and we need to update the module
name for consistency.

Test Plan: Terraform workflow

Reviewers: sneelakantan

Reviewed By: sneelakantan

Differential Revision: https://phorge.dev.yugabyte.com/D25668