[PLAT-9017]Send proper task failure errors after StateChangeConf.WaitForStateContext throws incorrect state
Summary:
Task Failures currently do not display the error message on the CLI running terraform. It displayed only the received state of the task and the target state.
This diff uses the new API introduced in `2.18.1.0-b68` to parse the list of failed subtasks and display the error on the command line.

The older versions and 2.19.0.0 would display a generic message to get the state of the Task failure from the YugabyteDB Anywhere UI Tasks Page.

Test Plan: Terraform workflow which produces an error displayed the correct message.

Reviewers: sneelakantan

Reviewed By: sneelakantan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D25931